svsm: Initialise gdbstub on startup and enter debugger on panics
Initialises the GDB stub module and hook it into the required exception
handlers to allow a remote debugger to inspect the state of the SVSM
module when execution is paused.

The default panic handler now triggers a debug breakpoint to allow a
remote gdb to inspect state when a panic occurs.

The debugger is only active for debug builds by default. When the gdb
stub feature is disabled (for release builds) placeholder functions are
used for all debugger actions.

Signed-off-by: Roy Hopkins <rhopkins@suse.de>