mm: fix SVSM memory overlap check
Previously the check was simply broken e.g. if the region started at
the same location that the SVSM memory starts the check wouldn't have
caught this. Also, previously, the memory after the SVSM memory region
was simply discarded.

This patch fixes the overlap check and makes sure that no memory is
wasted.

Signed-off-by: Tom Dohrmann <erbse.13@gmx.de>