stage1: Fix rounding error in kernel fs bin size
The size of the embedded filesystem image is rounded up on a 4 byte
boundary. The end address passed to stage 2 for the filesystem uses
this rounded value meaning that the code that unpacks the archive does
not know the exact size of the archive. When rounding occurs, the
unpacker wrongly assumes there is another file at the end and panics.

This patch retains the alignment but reports the exact archive size to
stage2.

Signed-off-by: Roy Hopkins <rhopkins@suse.de>