SVSM/protocols: Zero validates pages before giving guest access
Zero out a page when it is validated and before giving other VMPLs
access to it. This is necessary to prevent a possible HV attack:

Attack scenario:
  1) SVSM stores secrets in VMPL0 memory at GPA A
  2) HV invalidates GPA A and maps the SPA to GPA B, which is in the
     OS range of GPAs
  3) Guest OS asks SVSM to validate GPA B
  4) SVSM validates page and gives OS access
  5) OS can now read SVSM secrets from GPA B

The SVSM will not notice the attack until it tries to access GPA A
again. Prevent it by clearing every page before giving access to
other VMPLs.

Reported-by: Tom Lendacky <thomas.lendacky@amd.com>
Signed-off-by: Joerg Roedel <jroedel@suse.de>