svsm: Add support for tasks, task switching and scheduling
This adds basic support for creating tasks which equate to processes
or threads that can be scheduled on a vCPU. Each task is represented
by a structure that contains the relevant context, which at the moment
consists of a task stack, pagetable and stack register.

Tasks are added to a task list in TASKS which is implemented as an
RB tree which is keyed on the priority of the task: the lower the key,
the more likely the task will be switched to. There are two methods
for setting the priority; one based on CPU time using rdtsc and the
other (the default) the amount of times a task has been scheduled.

This patch adds all the infrastructure to support this but does not yet
integrate it into the boot processor or secondary APs, so current
functionality is unchanged.

Signed-off-by: Roy Hopkins <rhopkins@suse.de>