svsm/debug: Update gdbstub to support multiple tasks/threads
This patch makes the gdbstub aware of the task list in SVSM allowing
the developer to inspect the state and call stack of any stopped
threads regardless of which vCPU the task is running on.

Each task has its own pagetable so the gdbstub now uses its own
stack and switches to the task pagetable whenever a task context
is queried or modified.

Signed-off-by: Roy Hopkins <rhopkins@suse.de>