github/actions: allow clippy::bad_bit_mask
We get the following warning on the macro-expanded RMPFlags struct:

error: &-masking with zero
   --> src/sev/utils.rs:128:1
    |
128 | / bitflags::bitflags! {
129 | |     pub struct RMPFlags: u64 {
130 | |         #[allow(clippy::bad_bit_mask)]
131 | |         const VMPL0 = 0;
...   |
144 | |     }
145 | | }
    | |_^
    |
    = help: for further information visit https://rust-lang.github.io/rust-clippy/master/index.html#bad_bit_mask
    = note: this error originates in the macro `__impl_bitflags` which comes from the expansion of the macro `bitflags::bitflags` (in Nightly builds, run with -Z macro-backtrace for more info)

This happens because the VMPL0 flag is equal to the empty set of
flags. We want to keep this however because it improves readability.
Adding an individual #[allow(...)] statement on the structure does
not work due to macro expansion, so add an exception to the clippy
action.

Signed-off-by: Carlos López <carlos.lopez@suse.com>