acpi: use a Vec instead of a raw pointer in ACPITableBuffer
Instead of manually allocating a chunk of memory to copy the ACPI
table data from the firmware, use a Vec, which already keeps track of
the buffer size and is more erognomic.

This also allows us to remove the Drop implementation for the
structure, as the Vec will be deallocated automatically.

Signed-off-by: Carlos López <carlos.lopez@suse.com>