svsm/cpu: refactor X86Regs to separate from interrupt frame
The current X86Regs structure is only used by exception handlers and
interrupt routines and as such contains the interrupt routine stack
frame within the structure. This change separates that out to break
registers and other context into their respective categories to allow
the structures to be reused in other situations.

Signed-off-by: Roy Hopkins <rhopkins@suse.de>