svsm/task: Add support for creating tasks and task switching
This adds basic support for creating tasks which equate to processes
or threads that can be executed on a vCPU. Each task is represented
by a structure that contains the relevant context, which at the moment
includes a task stack, pagetable, stack pointer.

Tasks can be created with Task::create() providing an entry point
address. The task context is activated with Task::set_current(). The
task context remains in scope until it calls Task::set_current() itself
with a new context.

This patch adds all the infrastructure to support task switching but
does not yet integrate task creation or switching into the SVSM
kernel. This is in anticipation of a scheduler being introduced.

Signed-off-by: Roy Hopkins <rhopkins@suse.de>