acpi: construct ACPI tables from byte slices instead of raw pointers
Since the ACPI table buffer has a Vec that tracks the content size,
pass a slice of that vector to the ACPITable constructor instead of a
raw pointer.

This enables the use of the more idiomatic Vec::extend_from_slice()
to construct the inner content buffer for the ACPITable, as well as
reducing the amount of unsafe in the function.

Signed-off-by: Carlos López <carlos.lopez@suse.com>