acpi: prevent potential overflows in load_acpi_cpu_info()
Bounds check accesses to the MADT table buffer to prevent malformed
entries from causing out of bounds reads. Do this by introducing a new
method that safely casts the buffer contents as an arbitrary type,
verifying that the type itself fits within the buffer bounds.

While we are at it, reduce the scope of unsafe to dereferencing the
casted pointers.

Signed-off-by: Carlos López <carlos.lopez@suse.com>