task/schedule: Add schedule() to switch between tasks
This adds a public schedule() function that is used to yield
execution of the current task on a vCPU and switch to the
highest priority task on the per-cpu scheduler tree.

Creating new tasks now triggers the task to be allocated
to a vCPU and then calls schedule() so the task is given an
opportunity to run immediately. The new task can yield to
the next task by calling schedule().

Signed-off-by: Roy Hopkins <rhopkins@suse.de>