debug/gdbstub: Rework gdbstub to support multiple threads
The existing gdbstub assumes single thread operation. This
means that a developer using gdb is not able to determine
which vCPU is currently within context. Also, with the addition
of tasks to coco-svsm, the existing gdbstub does not allow
the debugger to determine which tasks are currently running
or stopped or which task the current thread of execution belongs
to.

This change reworks the gdbstub to query the task lists to
determine the state of each task and translates this to gdb
threads. Also, because each task maintains its own pagetable,
the gdbstub uses its own stack for the debugger and switches
pagetable when querying memory for each task. Therefore the
default stack is now not increased when the gdbstub feature
is enabled.

Signed-off-by: Roy Hopkins <rhopkins@suse.de>