debug/gdbstub,cpu/idt: Report page faults as SIGSEGV in gdbstub
The gdbstub currently triggers a breakpoint on a panic, which allows
the debugger to directly probe the state of the kernel when a panic
occurs. However, some of the context is lost before the panic handler
is invoked. Therefore this patch introduces direct handling of unhandled
page faults within the gdbstub to provide that additional context.

If the gdbstub feature is not enabled then these changes have no effect.

Signed-off-by: Roy Hopkins <rhopkins@suse.de>