mm/vm, fs: Add support for mapping vm to memory-backed files
For files that are backed by memory pages, which currently is all files
as we only support a ram filesystem, add support for creating
VirtualRanges that can map different portions of a file's physical pages
into virtual memory using a VMR.

This patch introduces a new struct named VMFileMapping which, given a
handle to a file, allows a portion of or the entire file into virtual
memory with requested permissions.

The VMFileMapping can then be inserted intoa VMR which will be used to
update the corresponding pagetable, giving access to the file contents
via a virtual memory pointer.

The virtual memory can be requested with write permission. With this
commit, the permission is granted but any writes are committed to the
original backing store. This will be fixed in a subsequent patch to
implement copy-on-write for the modified pages.

Signed-off-by: Roy Hopkins <rhopkins@suse.de>