greq: Add SnpGuestRequestDriver
Add a driver to send SNP_GUEST_REQUEST commands to the PSP. The command can be
any of the request or response command types defined in the SEV-SNP spec,
regardless if it's a regular or an extended command.

The send_regular_guest_request() and send_extended_guest_request() functions
can be used to send regular and extended commands, respectively.

guest_request_driver_init() is used to initialize the static driver instance.

Signed-off-by: Claudio Carvalho <cclaudio@linux.ibm.com>