task/tasks: Add support for launching tasks at CPL-3
The Task structure contains the 'user' field that optionally
provides information about the CPL-3 support of the task.
This commit adds the user_create() function which initialises
this field and allows the entry point of the task to be launched
at CPL-3.

In order to prevent the CPL-3 code from manipulating kernel
structures and execution flow, the task code that runs at CPL-3
does not share the task kernel stack. Instead a new user-mode
stack is created which is swapped in the transition from CPL-0 to
3 and back.

The task entry point is called from launch_user_entry() which runs
at CPL-0 and is responsible for performing the privilege level
switch as well as storing the kernel context for returning to CPL-0
via syscall. The syscall handler that performs this switch is already
present from a previous commit in task/syscall.rs.

Signed-off-by: Roy Hopkins <rhopkins@suse.de>