fw_meta: fix metadata table length calculation
The length field in the firmware metadata takes into account the size
of the length field itself plus the associated UUID. To get the size
of the contained tables, one must substract those two values. In
parse_fw_meta_data(), this was done incorrectly due to a lack of
parenthesis and operator precedence working from left to right.

For comparison, this is already done correctly in find_table():

    let len = orig_len - (mem::size_of::<Uuid>() + mem::size_of::<u16>());

Since the metadata is parsed backwards, the excessive computed length
would result in accesses to the memory before the firmware metadata.
This region would always be mapped, so no memory faults could end up
occuring.

Signed-off-by: Carlos López <carlos.lopez@suse.com>