address_space: add STACK_PAGES_GDB constant
Some new features may require increasing the stack size:
https://github.com/coconut-svsm/svsm/pull/69#discussion_r1282900603

To avoid problems when the gdb stub is enabled, it is best to define
STACK_PAGES as a constant plus extra pages depending on the features
(e.g. STACK_PAGES_GDB).

This way we have a single point to modify for the base of STACK_PAGES.

No functional changes.

Signed-off-by: Stefano Garzarella <sgarzare@redhat.com>