fw_meta: reduce raw pointer and unsafe usage during metadata parsing
Refactor parse_fw_meta_data() to avoid using unsafe and raw pointers
as much as possible. This mainly consists in properly typing the
metadata structures and going from raw pointers to typed references
or values as soon as possible. Since the firmware metadata comes from
the hypervisor, it can be considered untrusted in a confidential
computing model, so extra care is taken when using lengths and
offsets to index into memory.

Signed-off-by: Carlos López <carlos.lopez@suse.com>