cpu/vc: add CPUID handling in the #VC handler
There would be 2 ways to handle CPUID #VC exception: using the CPUID
table provided by the firmware (already mapped in memory in stage 2),
and forwarding the CPUID instruction to the hypervisor through a
VMEXIT.

Since the hyervisor is not trusted, we only handle CPUID with the
firmware table.

With this, since CPUID handling doesn't require the GHCB, stage 2 and
SVSM kernel have the same #VC handler.

Signed-off-by: Thomas Leroy <thomas.leroy@suse.com>