scripts: Build the svsm.bin from a container
Add scripts to build the project from a docker container. This can be
useful for multiple situations e.g., build it for a CI, quick build the
SVSM, or build it from a fresh environment.

$ ./scripts/docker/build.sh

The container will be delete after the SVSM is built, unless you pass -r
in the command line for it to reuse the coconut-build container, if it
exists. See the -h for more details.

$ ./scripts/docker/build.sh -r

Signed-off-by: Claudio Carvalho <cclaudio@linux.ibm.com>