build: use stable Rust toolchain
Switch to using the stable Rust toolchain instead of nightly. To do
this, we must not need to rebuild parts of the core and standard
library, which is an unstable feature. In order to do so, switch to a
supported build target (x86_64-unknown-none) instead of our own
(svsm-target.json).

Signed-off-by: Carlos López <carlos.lopez@suse.com>