SVSM/mm/pagetable: Implement struct PageTablePart
This struct represents a partial page-table and will be the base
struct for sharing sub-trees of page-tables among tasks.

Page-table freeing is also implemented on PageTableParts, as those may
be shared between many page-tables.

At the moment this means some code duplication with struct PageTable,
but that will be removed when the conversion is finished.

Signed-off-by: Joerg Roedel <jroedel@suse.de>