spinlock: remove unlock()
Remove Spinlock::unlock() because it bypasses the RAII lock tracking
logic of LockGuard.  This function is not used anywhere anyway.

The only way to lock the SpinLock is by acquiring a LockGuard instance
via lock() or try_lock().  LockGuard automatically releases the lock
when dropped.  There is no need for unlock(), it just allows messing up
the holder counting.

Signed-off-by: Oliver Steffen <osteffen@redhat.com>