SVSM: Increase boot-stack size
The 8 KiB of boot stack are not enough anymore after the switch to
stable rust. Especially in debug builds the stack usage is higher.
The maximum value I measured so far was 9608 bytes, which leads to
early boot crashes. Increase the boot stack size to 16 KiB to be on
the safe side.

Fixes: Github issue #90
Signed-off-by: Joerg Roedel <jroedel@suse.de>