SVSM/mm: Introduce Virtual Memory Mapping trait and data structure
Introduce struct VMM to represent single ranges of the virtual address
space. A VMM can be backed by different sources, which need to
implement the also added VirtualMemory trait.

The struct VMM is also prepared to be put into an RBTree. This will be
used to keep a data structure of existing mappings.

Signed-off-by: Joerg Roedel <jroedel@suse.de>