SVSM/mm/vm: Add Virtual Memory Range structure
Add a data structure for managing virtual memory ranges. It is named
struct VMR and manages struct VMM mappings within a given range of the
virtual address space.

Each struct VMR covers one or multiple entries of the top-level
page-table page. On x86-64 with 4-level paging this means the covered
size will always be a multiple of 512GB.

A struct VMR will manage an RBTree of the VMMs in that range and the
page-table sub-trees (PageTableParts) needed to map the VMR.

A top-level page-table can be generated by populating the
PageTableParts of several struct VMRs into it.

Signed-off-by: Joerg Roedel <jroedel@suse.de>