locking/rwlock: fix undefined behavior via RWLock::lock_read()
Fix an undefined behavior that could be triggered when acquiring a
read guard on a RWLock. This happened because a ReadLockGuard would
keep a mutable reference to the underlying shared data. Since one
might acquire multiple ReadLockGuards to the same data, this would
lead to undefined behavior under Rust's borrowing rules.

Fixes: b29acca1 ("SVSM/locking: Add RWLock implementation")
Reported-by: Tom Dohrmann <erbse.13@gmx.de>
Signed-off-by: Carlos López <carlos.lopez@suse.com>