mm/alloc: acquire Slab spinlock when freeing a pointer
Get access to the Slab corresponding to an allocation to be freed by
acquiring a lock first. Before, this could lead to undefined
behavior, as freeing an address would access the Slab mutably,
regardless of any concurrent users, leading to potential race
conditions.

Since the size of the allocation is provided via the layout parameter,
we can find the corresponding Slab via `SvsmAllocator::get_slab()`.

Partial-Bug: #107
Signed-off-by: Carlos López <carlos.lopez@suse.com>