locking: fix requirements for Sync & impl Send
The `Send` impl requires `Send` because RWLock the value will be
mutably accessible to another thread if the RWLock is sent to another
thread.
The `Sync` impl requires `Send` because another thread that has a
shared reference can acquire a mutable reference to the contained value
by write locking the RWLock. The `Sync` impl requires `Sync` because
it's possible for multiple threads to get a shared reference by read
locking the RWLock.

Signed-off-by: Tom Dohrmann <erbse.13@gmx.de>