mm/alloc: inline `remove_slab_page`
... into the only location where this function is called.

Before this patch, if the first page (i.e. `self.common.page`) was
free, we removed it with
`self.common.remove_slab_page(unsafe { &mut *last_page }, slab_page);`.
The problem with this was that we created two mutable references to
`self.common.page`, one through `last_page` and another implicitly
through `self.common`. This is obviously not allowed and it seems like
the as `*mut SlabPage` was added to stop the borrow checker from
complaining about this.
By inlining `remove_slab_page` we no longer need to create a reference
`self.common` and avoid needing two mutable references.

Fixes: #107
Signed-off-by: Tom Dohrmann <erbse.13@gmx.de>