mm/alloc: do not store Slab pointer in SlabPageInfo
We used to need to store a pointer to a Slab in the page information
in order to gain access to such slab it when freeing a pointer. Since
we receive the size of the allocation when such pointer, we can
retrieve the corresponding slab based on this information. Thus, we
no longer need to store a pointer to the slab.

While we are at it, remove the pointer from functions that used to
need it as an argument in order to store it.

Signed-off-by: Carlos López <carlos.lopez@suse.com>