acpi/tables: fix undefined behavior due to pointer misalignment
Fix an undefined behavior that would occur when the alignment
requirements for `core::slice::from_raw_parts()` would not be held.
The pointer passed to this function must be aligned so that each
member of the slice is itself aligned, which could not be
guaranteed.

Instead of creating a new slice of u32's, iterate over the individual
bytes in chunks of 4 and parse each u32 individually.

Fixes: #115
Fixes: 62f5c6fe "acpi: return ACPITable content as byte slice instead of raw pointer"
Signed-off-by: Carlos López <carlos.lopez@suse.com>