greq: Add SnpGuestRequestMsg and SnpGuestRequestExtData
These structures are used in the SNP_GUEST_REQUEST communication between the
guest and the PSP; their implementation follow the AMD SEV-SNP specification,
chapter 7.

The SnpGuestRequestMsg is used to carry a SNP_GUEST_REQUEST command or response
in the payload, which is encrypted using AES-256 GCM. This message can't be
tampered with by the hypervisor because only the PSP and the guest have access
to the key to decrypt the payload.

An extended SNP_GUEST_REQUEST command also requests data from the hypervisor;
in this case, the SnpGuestRequestExtData is also provided. The hypervisor will
use it to store the requested data.

Signed-off-by: Claudio Carvalho <cclaudio@linux.ibm.com>