tests/mm: use RAII guard to drop test root memory
Replace setup_test_root_mem() and destroy_test_root_mem() by a new
guard-like struct that, when acquired, grabs a global lock and
initializes the test memory, and when dropped it destroys the memory
before releasing the lock.

This simplifies some tests by removing the need to manually destroy
the global memory. Moreover, guards on ROOT_MEM do not need to be
dropped manually either, since they are always acquired after the
memory lock, and Rust drops objects in reverse order of acquisition.

Signed-off-by: Carlos López <carlos.lopez@suse.com>