mm/alloc: simplify SvsmAllocator::get_slab() logic
The SvsmAllocator manages a list of slabs of different sizes. Since
these sizes of powers of 2, we can exploit this fact to find the
adequate slab for a given allocation size:

  * First, round the allocation size to the next power of 2 to find
    the required slab size.
  * Next, count the number of trailing zeros, substract the trailing
    zeros of the smallest slab size, and use that as an index into an
    array of slabs.

For example, 128 has 7 trailing zeros, which is 2 more than the
smallest slab size (32). Thus, the 128-byte slab is at index 2.

This simplifies the logic, removing several conditional branches.

Signed-off-by: Carlos López <carlos.lopez@suse.com>