percpu: Assign apic_id in new()
PerCpu.new() currently initialises the apic_id to 0. This is then
reassigned to the correct apic_id after creation. The only place new()
is called is directly from PerCpu.alloc(). Therefore this commit adds
the apic_id as a parameter to new() and makes it private.

This is in preparation adding a runqueue to the PerCpu structure which
will require access to the apic_id on creation.

Signed-off-by: Roy Hopkins <rhopkins@suse.de>