task, percpu: Use VMR to manage the task pagetable and VMKernelStack to allocate the task kernel stack
Now that virtual memory ranges and page table entries can be managed
using the new `VMR` and associated structs, update the Task structure
implementation to manage the task pagetable using these new
capabilities.

The task structure now includes a `VMR` instance that defines the
virtual address range for the task. The kernel stack for the task is
created using `VMKernelStack` and added to the task VMR. This is applied
to the task pagetable during creation.

`PerCpu` already includes a per cpu VMR. This needs to be applied to the
task when it is context switched onto the CPU. Currently this is done
directly by copying the PerCpu level 3 index. This patch adss the
function `populate_page_table()` to PerCpu which appiles the PerCpu VMR
to the pagetable instead.

Signed-off-by: Roy Hopkins <rhopkins@suse.de>