task/tasks: Add an optional hook that can be called after task switch
This adds a new field to the `Task` structure that optionally can be
set to a function that is called immediately after switching to the
pagetable and stack of the incoming task during a task switch.

This can be used to perform post-switch tasks such as releasing locks
that are held during the task switch.

Signed-off-by: Roy Hopkins <rhopkins@suse.de>