mm: Add VMRMapping to manage a temporary mapping with a VMR
It is commonly required to temporarily map a Mapping range into a VMR.
For example, when setting up a kernel stack for a new task, the task
stack frame needs to be initialized by the parent task. In this case the
kernel stack mapping needs to be added to the task or PerCpu VMR to get
a virtual address which can be used to populate the initial stack.

This commit adds a new struct, VMRMapping which takes a mapping and
inserts it into a provided VMR instance. The mapping is removed from the
instance when the VMRMapping is dropped.

Signed-off-by: Roy Hopkins <rhopkins@suse.de>