mapping: Allow pt_flags within VirtualMapping to be specified per page
Currently a VirtualMapping instance returns the required permissions as
a single set of flags for all pagetable entries across the mapping. In
order to support copy-on-write behaviour in future structs that
implement the VirtualMapping trait we need to be able to specify the
flags with page granularity.

This commit adds a parameter to the pt_flags() function that requires
the offset into the mapping to be specified. The function then returns
the flags of the page that contains the offset.

No existing VirtualMapping implementation use the offset so the
parameter is currently ignored.

Signed-off-by: Roy Hopkins <rhopkins@suse.de>