mapping: Add page fault handler to virtual memory ranges
In order to implement lazy allocation and copy-on-write behaviour within
virtual memory ranges, the associated structures need to be notified
when page faults occur within virtual memory ranges giving the
opportunity to allow changes to be made to the memory configuration to
satisify the fault condition.

This commit adds functions to VMR and the VirtualMapping traits and
structures to allow the interrupt handler to forward page faults and
allow them to be handled within the specific mapping implementations.

A VirtualMapping will return either a VMPageFaultResolution that
requests a particular change to the pagetable, or an SvsmError which
means that the page fault should propogate to the next handler.

Signed-off-by: Roy Hopkins <rhopkins@suse.de>