rawalloc: Allow for late allocation of pages within RawAllocMapping
Currently RawAllocMapping can allocate either none or all pages of the
size when created via RawAllocMapping::new(). This commit extends the
implementation to allow selective pages to be allocated, keeping track
of which pages have currently been allocated.

The vector of pages has been modified to contain optional PageRefs with
the vector size being fixed in new() instead of alloc_pages().
alloc_pages() still allocates all pages but an additional function is
provided that allows individual pages to be allocated. In addition, the
presence of a page can be checked the new function, present().

This modified functionality allows the RawAllocMapping to be used
as currently with no functional changes but also allows sparse memory
allocation. This is useful for copy-on-write implementations where a
pages is only allocated when a write page fault occurs, allocating a
page and copying from the read-only source.

Signed-off-by: Roy Hopkins <rhopkins@suse.de>