mapping/file_mapping: Implement copy-on-write for writeable file mapped vm regions
Currently, a writeable VMFileMapping will copy the entire contents of
the region from the source pages into a read/write copy. This is
inefficient in cases where the mapping may be quite large or only
sparsely accessed.

This commit adds page fault handling to VMFileMapping. The original file
pages are mapped as read-only and when a write occurs the page fault is
handled and a copy taken of the page where the fault occurred. The copy
is stored in a sparse vector in RawAllocMapping. The VMR will update the
pagetable for the virtual range to map the new page with the correct
access flags.

Signed-off-by: Roy Hopkins <rhopkins@suse.de>