mapping/file_mapping: Add unit tests for VMFileMapping
This commit contains a small refactor where the code that copies a page
of file data to a writeable page is moved into a new function so it can
be replaced by a test implementation during unit test execution.

Tests have been added that check the mapping returned by VMFileMapping
matches the correct offsets within the source file and that the
copy-on-write logic with page faults is handled correctly.

Signed-off-by: Roy Hopkins <rhopkins@suse.de>