mm/alloc: introduce MemoryRegion::get_pfn()
Currently, there are two different paths to get the information for a
specific page:

1. Passing a pfn to `MemoryRegion::read_page_info()`, which will panic
on an invalid pfn (via `MemoryRegion::check_pfn()`).

2. Passing a `VirtAddr` to `MemoryRegion::get_page_info()`, which will
return an error on an invalid address, and then will call
`MemoryRegion::read_page_info()` by computing the correct pfn for the
address.

This is a bit confusing and is redundant in some cases, as some of the
callers of `MemoryRegion::get_page_info()` also need the pfn
afterwards, which means the pfn is computed twice.

To simplify this, decouple the `VirtAddr` to pfn conversion from the
information reading by introducing `MemoryRegion::get_pfn()`. The new
method will perform all the necessary checks on the `VirtAddr`, and
will return a valid pfn that can be used to retrieve the page
information. This allows the removal of
`MemoryRegion::get_page_info()`.

In theory, this could also allow us to remmove
`MemoryRegion::check_pfn()`, as all the pfns are coming from either
the new function or the page freelist. However, we leave it in to
catch corruption issues in the freelist.

Signed-off-by: Carlos López <carlos.lopez@suse.com>