mm/alloc: return a pointer from MemoryRegion::page_info_virt_addr()
`MemoryRegion::page_info_virt_addr()` is used to read or write page
information for a provided pfn. The returned `VirtAddr` is always
casted to a pointer to a `PageStorageType` by the callers, so return
that type directly. The arithmetic is also simplified, since the
primitive pointer type has an `add()` method that takes into account
the size of the generic type `T` of the pointer.

To make things more clear, rename `page_info_virt_addr()` to
`page_info_ptr()`. Mark it unsafe as well, as the addition of the pfn
offset with overflow will cause undefined behavior, according to the
documentation of the `add()` method.

Signed-off-by: Carlos López <carlos.lopez@suse.com>