mm/alloc: rename Page enum and variants
Rename the Page enum to PageInfo, and remove the "Page" suffix of
variants' names that have it to improve readability. This also fixes
the following warnings that will appear on a future (e.g. nightly)
version of clippy:

warning: variant name ends with the enum's name
   --> src/mm/alloc.rs:197:5
    |
197 |     SlabPage(SlabPageInfo),
    |     ^^^^^^^^^^^^^^^^^^^^^^
    |
    = help: for further information visit https://rust-lang.github.io/rust-clippy/master/index.html#enum_variant_names
    = note: `#[warn(clippy::enum_variant_names)]` on by default

warning: variant name ends with the enum's name
   --> src/mm/alloc.rs:198:5
    |
198 |     CompoundPage(CompoundInfo),
    |     ^^^^^^^^^^^^^^^^^^^^^^^^^^
    |
    = help: for further information visit https://rust-lang.github.io/rust-clippy/master/index.html#enum_variant_names

warning: variant name ends with the enum's name
   --> src/mm/alloc.rs:199:5
    |
199 |     FilePage(FileInfo),
    |     ^^^^^^^^^^^^^^^^^^
    |
    = help: for further information visit https://rust-lang.github.io/rust-clippy/master/index.html#enum_variant_names

Signed-off-by: Carlos López <carlos.lopez@suse.com>