greq/msg: fix reachable assertions due to insufficient allocation layout
From the docs [0]:

> An array of [T; N] has a size of size_of::<T>() * N and the same
> alignment of T

SnpGuestRequestExtData is an array of u8s, while SnpGuestRequestMsg
contains SnpGuestRequestMsgHdr, a packed struct of 96 bytes, followed
by a u8 array.

This means that when allocating these structures via their
boxed_new() methods an assertion might trigger if the address is not
page aligned, which the allocator may legally return.

Fix this by setting an explicit alignment. Both structures can afford
not being packed: SnpGuestRequestExtData contains a contiguous byte
array, so there will not be padding; SnpGuestRequestMsg contains only
two fields, of which the first one ends naturally at an 8-byte
boundary, and the second one has a required alignment of 1 (a u8
array), which means there will never be padding.

While we are at it, add two tests to verify that the assertions are
not triggered.

[0] https://doc.rust-lang.org/reference/type-layout.html#array-layout

Fixes: 9211b4fa ("greq: Add SnpGuestRequestMsg and SnpGuestRequestExtData")
Signed-off-by: Carlos López <carlos.lopez@suse.com>