mm/alloc/tests: enable TestRootMem for tests inside the SVSM
Tests relying on the memory allocator need to set up a `TestRootMem`
when running in userspace to emulate a regular boot of the SVSM where
the heap has been set up. `TestRootMem` simply requests a chunk of
memory from the standard Rust allocator and initializes the SVSM root
memory.

However, when running the tests inside the SVSM, root memory is
already set up, as the kernel has booted normally. Thus, do not
attempt to reinitialize root memory and simply use the regular
allocator. We still use a global lock to keep a consistent API.

Do this by gating `TestRootMem::setup()` depending on whether
cfg(test_inside_svsm) is set. This allows tests to use `TestRootMem`
transparently.

Signed-off-by: Carlos López <carlos.lopez@suse.com>