fs/filesystem/tests: introduce TestFileSystemGuard
Introduce a new guard-like struct to handle initialization and
destruction of a filesystem for testing. This structure allows writing
tests for the filesystem transparently, as it will act as a no-op
when running tests inside the SVSM. This is need because in the latter
case, the filesystem is already managed by the SVSM kernel.

Signed-off-by: Carlos López <carlos.lopez@suse.com>