igvm_params: Replace launch_fw with firmware range in IGVM parameter block
The IgvmParamBlock currently contains a flag that determines whether the
guest firmware (OVMF) should be launched or not by the SVSM kernel.
However, in order to launch the firmware embedded in an IGVM file the
SVSM needs to know the GPA at which the firmware is loaded and the
number of pages in order to set the correct VMPL permissions on the
pages.

For the non-IGVM case, this comes from fw_cfg but this is not populated
when using IGVM files. So this commit replaces the launch_fw flag with a
start GPA and page range for the firmware region.

Signed-off-by: Roy Hopkins <roy.hopkins@suse.com>